+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## 25.1.0
 
 - initial version
+
+## 25.1.1
+- Ignore symlinks
+- Dehydrate local file when it was changed in the cloud
+
+## 25.1.2
+- Change app display name